Trade Execution Snark (#26)
This is the main snark that validates a given solution is executed properly. It includes test case for valid instances as well as test cases for all possible invalid solutions I could think of.

At the moment it is expecting all prices and volumes to match exactly, an assumption which might have to be relaxed in the future.