Adding rounding to trade execution snark (#27)
Before, the trade execution snark is expecting values to match exactly. In particular, the relations between buy/sell volume and prices, the total sum of tokens bought/sold as well as the trader surplus where exact equalities.

As sketched out in #14 this limits the region of valid solutions drastically. Therefore, we are introducing an epsilon (and unit tests to show which kind of values are now variable) to allow ≈ instead of =.